+++ src/main/resources/application.properties
@@ -13,7 +13,7 @@ spring.servlet.session.tracking-modes=cookie
 spring.jpa.database=mysql
 spring.jpa.database-platform=org.hibernate.dialect.MySQL5InnoDBDialect
 spring.jpa.generate-ddl=false
-spring.jpa.hibernate.ddl-auto=none
+spring.jpa.hibernate.ddl-auto=create
 
 # sql setting
 spring.jpa.open-in-view=false
@@ -34,8 +34,3 @@ logging.level.org.apache.coyote.http11=debug
 
 # URL jsessionid exposure prevention
 server.servlet.session.tracking-modes=cookie
-
-
-
-
-
