+++ src/test/java/SilkLoad/DemoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package SilkLoad;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
