+++ src/test/java/SilkLoad/repository/ProductImageRepositoryTest.java
@@ -1,0 +1,7 @@
+package SilkLoad.repository;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class ProductImageRepositoryTest {
+
+}
